Make RecipeBookManagerMixin clientside (#1532)
This class isn't marked as OnlyIn(Dist.CLIENT), but it is in NeoForge's client package. It also references many classes which are marked as client, so loading this mixin would cause issues on a server. This is normally not noticeable, but if the mixin is forcibly loaded via a mixin audit it will crash.